refactor(scores): extract event score-facet shaping into a sibling module
Move the pure in-memory shaping for the combined score-name facet scan
(row types, the empty-options constant, and the group/rank/cap helpers)
out of the repository file into scoreEventFacets.ts. The repository keeps
only the ClickHouse round-trip and the score-config value merge.
